Replace assertions with exception logging and error handling
Assertions where causing the program to behave unpredictably, and in many instances to stop working.

Here assertions where removed in favor of error logging to console, and handling the error by gracefully cancelling the action.

The user is alerted to the error by a quick red flash around the edges of the window.

Handling of these errors has made the program more stable, not less